Seperate skidsParser from skids definitions.
